+++ package.json
@@ -2,14 +2,19 @@
     "name": "angular2-material-dashboard-pro",
     "version": "0.0.0",
     "license": "MIT",
+    "engines": {
+        "node": "10.5.0",
+        "npm": "6.1.0"
+    },
     "scripts": {
         "ng": "ng",
-        "start": "ng serve",
+        "start": "npm run start --prefix server",
         "build": "ng build --prod",
         "test": "ng test",
         "lint": "ng lint",
         "e2e": "ng e2e",
-        "serve": "ng serve --open"
+        "serve": "ng serve --open",
+        "heroky-postbuild": "npm run build"
     },
     "private": true,
     "dependencies": {
