+++ server/config/index.js
@@ -1,0 +1,1 @@
+module.exports = process.env.NODE_ENV === 'production' ? require('./prod') : require('./dev');
