+++ server/config/prod.js
@@ -1,0 +1,5 @@
+module.exports = {
+    DB_URI: "mongodb+srv://demonodejs:k9mMg8r9cWij4rfb@cluster0-2zokl.mongodb.net/test?retryWrites=true",
+    SECRET: 'helloworld!',
+    STRIPE_SK: "sk_test_ig5LvYbPJLcYsccyKRb6HujA00UEy7Ak8D"
+}
