+++ server/index.js
@@ -2,17 +2,22 @@ const express = require('express');
 const app = express();
 const PORT = process.env.PORT || 3001;
 const mongoose = require('mongoose');
-const config = require('./config/dev');
+const config = require('./config');
 const FakeDb = require('./fake-db');
 const bodyParser = require('body-parser');
+const path = require('path');
+
 const rentalRouters = require('./routes/rental');
 const userRouters = require('./routes/user');
 const bookingRouters = require('./routes/booking');
 const paymentRoutes = require('./routes/payment');
 
 mongoose.connect(config.DB_URI, { useNewUrlParser: true }).then(() => {
-    const fakedb = new FakeDb();
-    // fakedb.seedDb();
+    if (process.env.NODE_ENV !== 'production') {
+        const fakedb = new FakeDb();
+        // fakedb.seedDb();
+    }
+    
 });
 
 app.use(bodyParser.json());
@@ -27,8 +32,18 @@ app.use('/api/v1/user', userRouters);
 app.use('/api/v1/bookings', bookingRouters);
 app.use('/api/v1/payments', paymentRoutes);
 
+if (process.env.NODE_ENV === 'production') {
+    const appPath = path.join(__dirname, '..', 'dist');
+    app.use(express.static(appPath));
+
+    app.get('*', (req, res) => {
+        res.sendFile(path.resolve(appPath, 'index.html'));
+    });
+}
+
 app.listen(PORT, () => {
     console.log(`port ${PORT} running`);
+    console.log(process.env.NODE_ENV);
 });
 
 // NOTE: Relationship 1-N
