+++ src/environments/environment.prod.ts
@@ -1,3 +1,4 @@
 export const environment = {
-  production: true
+  production: true,
+  STRIPE_PK: "pk_test_zw42Y66nR4E9Gn5OnFl7OkBa006x5joN11"
 };
