+++ src/index.html
@@ -3,7 +3,7 @@
 
 <head>
     <meta charset="utf-8">
-    <title>Angular2 Material Dashboard Pro</title>
+    <title>Booking Rental Example</title>
     <base href="/">
 
     <meta name="viewport" content="width=device-width, initial-scale=1">
